+++ lib/src/release_updater_provider_http.dart
@@ -15,25 +15,7 @@ class ReleaseProviderHttp extends ReleaseProvider {
 
   final String releasesBundleFileFormat;
   static const String defaultReleasesBundleFileFormat =
-      '%NAME%-%VER%%[-]PLATFORM%.zip';
-
-  static String formatBundleFile(
-      String fileFormat, String name, Version version,
-      [String? platform]) {
-    var file = _replaceMark(fileFormat, 'NAME', name);
-    file = _replaceMark(file, 'VER', version.toString());
-    file = _replaceMark(file, 'PLATFORM', platform);
-    return file;
-  }
-
-  static String _replaceMark(String s, String mark, String? value) {
-    var regExp = RegExp(r'%(?:\[(.*?)\])?(' + mark + r')%');
-
-    return s.replaceAllMapped(regExp, (m) {
-      var prev = m.group(1) ?? '';
-      return value != null ? '$prev$value' : '';
-    });
-  }
+      ReleaseBundle.defaultReleasesBundleFileFormat;
 
   ReleaseProviderHttp.withClient(this.httpClient,
       {this.releasesFile = defaultReleasesFile,
@@ -66,7 +48,7 @@ class ReleaseProviderHttp extends ReleaseProvider {
   @override
   Future<ReleaseBundle?> getReleaseBundle(String name, Version targetVersion,
       [String? platform]) async {
-    var file = formatBundleFile(
+    var file = ReleaseBundle.formatReleaseBundleFile(
         releasesBundleFileFormat, name, targetVersion, platform);
     var response = await httpClient.get(file);