+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Improved commands logging.
 - Fix passing of release bundle to finalize commands.
 - Handler `URL` request errors.
+- `ReleasePackerCommandUploadReleaseBundle``: fix parameter `file` format.
 
 ## 1.0.11